+++ Kernel/cpan-lib/Text/CSV_PP.pm
@@ -11,7 +11,7 @@ use strict;
 use vars qw($VERSION);
 use Carp ();
 
-$VERSION = '1.29';
+$VERSION = '1.31';
 
 sub PV  { 0 }
 sub IV  { 1 }
@@ -19,6 +19,7 @@ sub NV  { 2 }
 
 sub IS_QUOTED () { 0x0001; }
 sub IS_BINARY () { 0x0002; }
+sub IS_MISSING () { 0x0010; }
 
 
 my $ERRORS = {
@@ -60,6 +61,8 @@ my $ERRORS = {
         3006 => "EHR - bind_columns () did not pass enough refs for parsed fields",
         3007 => "EHR - bind_columns needs refs to writable scalars",
         3008 => "EHR - unexpected error in bound fields",
+        3009 => "EHR - print_hr () called before column_names ()",
+        3010 => "EHR - print_hr () called with invalid arguments",
 
         0    => "",
 };
@@ -78,6 +81,7 @@ my %def_attr = (
     keep_meta_info      => 0,
     allow_loose_quotes  => 0,
     allow_loose_escapes => 0,
+    allow_unquoted_escape => 0,
     allow_whitespace    => 0,
     chomp_verbatim      => 0,
     types               => undef,
@@ -87,8 +91,11 @@ my %def_attr = (
     auto_diag           => 0,
     quote_space         => 1,
     quote_null          => 1,
+    quote_binary        => 1,
+    diag_verbose        => 0,
 
     _EOF                => 0,
+    _RECNO              => 0,
     _STATUS             => undef,
     _FIELDS             => undef,
     _FFLAGS             => undef,
@@ -106,10 +113,12 @@ BEGIN {
         $INC{'bytes.pm'} = 1 unless $INC{'bytes.pm'}; # dummy
         no strict 'refs';
         *{"utf8::is_utf8"} = sub { 0; };
+        *{"utf8::decode"}  = sub { };
     }
     elsif ( $] < 5.008 ) {
         no strict 'refs';
         *{"utf8::is_utf8"} = sub { 0; };
+        *{"utf8::decode"}  = sub { };
     }
     elsif ( !defined &utf8::is_utf8 ) {
        require Encode;
@@ -244,6 +253,11 @@ sub error_diag {
 
     return $context ? @diag : $diagobj;
 }
+
+sub record_number {
+    return shift->{_RECNO};
+} 
+
 ################################################################################
 # string
 ################################################################################
@@ -273,8 +287,8 @@ sub _combine {
     $self->{_STRING}      = '';
     $self->{_STATUS}      = 0;
 
-    my ($always_quote, $binary, $quot, $sep, $esc, $empty_is_undef, $quote_space, $quote_null)
-            = @{$self}{qw/always_quote binary quote_char sep_char escape_char empty_is_undef quote_space quote_null/};
+    my ($always_quote, $binary, $quot, $sep, $esc, $empty_is_undef, $quote_space, $quote_null, $quote_binary )
+            = @{$self}{qw/always_quote binary quote_char sep_char escape_char empty_is_undef quote_space quote_null quote_binary/};
 
     if(!defined $quot){ $quot = ''; }
 
@@ -312,7 +326,7 @@ sub _combine {
 
         if( $binary and $quote_null ){
             use bytes;
-            $must_be_quoted++ if ( $column =~ s/\0/${esc}0/g || $column =~ /[\x00-\x1f\x7f-\xa0]/ );
+            $must_be_quoted++ if ( $column =~ s/\0/${esc}0/g || ($quote_binary && $column =~ /[\x00-\x1f\x7f-\xa0]/) );
         }
 
         if($always_quote or $must_be_quoted){
@@ -340,9 +354,9 @@ sub _parse {
 
     return 0 if(!defined $line);
 
-    my ($binary, $quot, $sep, $esc, $types, $keep_meta_info, $allow_whitespace, $eol, $blank_is_undef, $empty_is_undef)
+    my ($binary, $quot, $sep, $esc, $types, $keep_meta_info, $allow_whitespace, $eol, $blank_is_undef, $empty_is_undef, $unquot_esc)
          = @{$self}{
-            qw/binary quote_char sep_char escape_char types keep_meta_info allow_whitespace eol blank_is_undef empty_is_undef/
+            qw/binary quote_char sep_char escape_char types keep_meta_info allow_whitespace eol blank_is_undef empty_is_undef allow_unquoted_escape/
            };
 
     $sep  = ',' unless (defined $sep);
@@ -357,14 +371,18 @@ sub _parse {
     my $re_split       = $self->{_re_split}->{$quot}->{$esc}->{$sep} ||= _make_regexp_split_column($esc, $quot, $sep);
     my $re_quoted       = $self->{_re_quoted}->{$quot}               ||= qr/^\Q$quot\E(.*)\Q$quot\E$/s;
     my $re_in_quot_esp1 = $self->{_re_in_quot_esp1}->{$esc}          ||= qr/\Q$esc\E(.)/;
-    my $re_in_quot_esp2 = $self->{_re_in_quot_esp2}->{$quot}->{$esc} ||= qr/[\Q$quot$esc\E0]/;
+    my $re_in_quot_esp2 = $self->{_re_in_quot_esp2}->{$quot}->{$esc} ||= qr/[\Q$quot$esc$sep\E0]/;
     my $re_quot_char    = $self->{_re_quot_char}->{$quot}            ||= qr/\Q$quot\E/;
-    my $re_esc          = $self->{_re_esc}->{$quot}->{$esc}          ||= qr/\Q$esc\E(\Q$quot\E|\Q$esc\E|0)/;
+    my $re_esc          = $self->{_re_esc}->{$quot}->{$esc}          ||= qr/\Q$esc\E(\Q$quot\E|\Q$esc\E|\Q$sep\E|0)/;
     my $re_invalid_quot = $self->{_re_invalid_quot}->{$quot}->{$esc} ||= qr/^$re_quot_char|[^\Q$re_esc\E]$re_quot_char/;
 
     if ($allow_whitespace) {
         $re_split = $self->{_re_split_allow_sp}->{$quot}->{$esc}->{$sep}
                      ||= _make_regexp_split_column_allow_sp($esc, $quot, $sep);
+    }
+    if ($unquot_esc) {
+        $re_split = $self->{_re_split_allow_unqout_esc}->{$quot}->{$esc}->{$sep}
+                     ||= _make_regexp_split_column_allow_unqout_esc($esc, $quot, $sep);
     }
 
     my $palatable = 1;
@@ -445,7 +463,8 @@ sub _parse {
                         $palatable = 0;
                         last;
                     }
-                    else {
+
+                    unless ($self->{allow_loose_quotes}) {
                         $col =~ s/\Q$esc\E(.)/$1/g;
                     }
                 }
@@ -522,10 +541,20 @@ sub _parse {
                 $col = undef;
             }
 
+            if ( $unquot_esc ) {
+                $col =~ s/\Q$esc\E(.)/$1/g;
+            }
+
+        }
+
+        utf8::encode($col) if $utf8;
+        if ( defined $col && _is_valid_utf8($col) ) {
+            utf8::decode($col);
         }
 
         push @part,$col;
         push @{$meta_flag}, $flag if ($keep_meta_info);
+        $self->{ _RECNO }++;
 
         $i++;
     }
@@ -552,10 +581,35 @@ sub _make_regexp_split_column {
         return qr/([^\Q$sep\E]*)\Q$sep\E/s;
     }
 
-   qr/(
+   return qr/(
         \Q$quot\E
             [^\Q$quot$esc\E]*(?:\Q$esc\E[\Q$quot$esc\E0][^\Q$quot$esc\E]*)*
         \Q$quot\E
+        | # or
+        \Q$quot\E
+            (?:\Q$esc\E[\Q$quot$esc$sep\E0]|[^\Q$quot$esc$sep\E])*
+        \Q$quot\E
+        | # or
+        [^\Q$sep\E]*
+       )
+       \Q$sep\E
+    /xs;
+}
+
+
+sub _make_regexp_split_column_allow_unqout_esc {
+    my ($esc, $quot, $sep) = @_;
+
+   return qr/(
+        \Q$quot\E
+            [^\Q$quot$esc\E]*(?:\Q$esc\E[\Q$quot$esc\E0][^\Q$quot$esc\E]*)*
+        \Q$quot\E
+        | # or
+        \Q$quot\E
+            (?:\Q$esc\E[\Q$quot$esc$sep\E0]|[^\Q$quot$esc$sep\E])*
+        \Q$quot\E
+        | # or
+            (?:\Q$esc\E[\Q$quot$esc$sep\E0]|[^\Q$quot$esc$sep\E])*
         | # or
         [^\Q$sep\E]*
        )
@@ -581,7 +635,7 @@ sub _make_regexp_split_column_allow_sp {
     qr/$ws*
        (
         \Q$quot\E
-            [^\Q$quot$esc\E]*(?:\Q$esc\E[\Q$quot$esc\E0][^\Q$quot$esc\E]*)*
+            [^\Q$quot$esc\E]*(?:\Q$esc\E[\Q$quot$esc$sep\E0][^\Q$quot$esc\E]*)*
         \Q$quot\E
         | # or
         [^\Q$sep\E]*?
@@ -606,6 +660,13 @@ sub print {
     local $\ = '';
 
     $io->print( $self->_string ) or $self->_set_error_diag(2200);
+}
+
+sub print_hr {
+    my ($self, $io, $hr) = @_;
+    $self->{_COLUMN_NAMES} or $self->_set_error_diag(3009);
+    ref $hr eq "HASH"      or $self->_set_error_diag(3010);
+    $self->print ($io, [ map { $hr->{$_} } $self->column_names ]);
 }
 ################################################################################
 # getline
@@ -639,7 +700,10 @@ sub getline {
         LOOP: {
             my $is_continued   = scalar(my @list = $line =~ /$re/g) % 2; # if line is valid, quot is even
 
-            if ( $line =~ /${re}0/ ) { # null suspicion case
+            if ( $self->{allow_loose_quotes } ) {
+                $is_continued = 0;
+            }
+            elsif ( $line =~ /${re}0/ ) { # null suspicion case
                 $is_continued = $line =~ qr/
                     ^
                     (
@@ -704,7 +768,6 @@ sub _return_getline_result {
 
     return [];
 }
-
 ################################################################################
 # getline_all
 ################################################################################
@@ -749,6 +812,10 @@ sub getline_hr {
     }
 
     my $fr = $self->getline( $io ) or return undef;
+
+    if ( ref $self->{_FFLAGS} ) {
+        $self->{_FFLAGS}[$_] = IS_MISSING for ($#{$fr} + 1) .. $#{$self->{_COLUMN_NAMES}};
+    }
 
     @hr{ @{ $self->{_COLUMN_NAMES} } } = @$fr;
 
@@ -857,6 +924,13 @@ sub is_binary {
     return if( $_[1] =~ /\D/ or $_[1] < 0 or  $_[1] > $#{ $_[0]->{_FFLAGS} } );
     $_[0]->{_FFLAGS}->[$_[1]] & IS_BINARY ? 1 : 0;
 }
+
+sub is_missing {
+    my ($self, $idx, $val) = @_;
+    ref $self->{_FFLAGS} &&
+            $idx >= 0 && $idx < @{$self->{_FFLAGS}} or return;
+    $self->{_FFLAGS}[$idx] & IS_MISSING ? 1 : 0;
+}
 ################################################################################
 # _check_type
 #  take an arg as scalar referrence.
@@ -895,7 +969,8 @@ sub _set_error_diag {
 
 BEGIN {
     for my $method ( qw/always_quote binary keep_meta_info allow_loose_quotes allow_loose_escapes
-                            verbatim blank_is_undef empty_is_undef auto_diag quote_space quote_null/ ) {
+                            verbatim blank_is_undef empty_is_undef quote_space quote_null
+                            quote_binary allow_unquoted_escape/ ) {
         eval qq|
             sub $method {
                 \$_[0]->{$method} = defined \$_[1] ? \$_[1] : 0 if (\@_ > 1);
@@ -971,6 +1046,41 @@ sub SetDiag {
     Carp::croak( $_[0]->error_diag . '' );
 }
 
+sub auto_diag {
+    my $self = shift;
+    if (@_) {
+        my $v = shift;
+        !defined $v || $v eq "" and $v = 0;
+        $v =~ m/^[0-9]/ or $v = $v ? 1 : 0; # default for true/false
+        $self->{auto_diag} = $v;
+    }
+    $self->{auto_diag};
+}
+
+sub diag_verbose {
+    my $self = shift;
+    if (@_) {
+        my $v = shift;
+        !defined $v || $v eq "" and $v = 0;
+        $v =~ m/^[0-9]/ or $v = $v ? 1 : 0; # default for true/false
+        $self->{diag_verbose} = $v;
+    }
+    $self->{diag_verbose};
+}
+
+sub _is_valid_utf8 {
+    return ( $_[0] =~ /^(?:
+         [\x00-\x7F]
+        |[\xC2-\xDF][\x80-\xBF]
+        |[\xE0][\xA0-\xBF][\x80-\xBF]
+        |[\xE1-\xEC][\x80-\xBF][\x80-\xBF]
+        |[\xED][\x80-\x9F][\x80-\xBF]
+        |[\xEE-\xEF][\x80-\xBF][\x80-\xBF]
+        |[\xF0][\x90-\xBF][\x80-\xBF][\x80-\xBF]
+        |[\xF1-\xF3][\x80-\xBF][\x80-\xBF][\x80-\xBF]
+        |[\xF4][\x80-\x8F][\x80-\xBF][\x80-\xBF]
+    )+$/x )  ? 1 : 0;
+}
 ################################################################################
 package Text::CSV::ErrorDiag;
 
@@ -1042,9 +1152,10 @@ is a XS module and Text::CSV_PP is a Puer Perl one.
 
 =head1 VERSION
 
-    1.29
+    1.31
 
-This module is compatible with Text::CSV_XS B<0.80> and later.
+This module is compatible with Text::CSV_XS B<0.99>.
+(except for diag_verbose and allow_unquoted_escape)
 
 =head2 Unicode (UTF8)
 
@@ -1764,7 +1875,7 @@ Text::CSV was written by E<lt>alan[at]mfgrtl.comE<gt>.
 
 =head1 COPYRIGHT AND LICENSE
 
-Copyright 2005-2010 by Makamaka Hannyaharamitu, E<lt>makamaka[at]cpan.orgE<gt>
+Copyright 2005-2013 by Makamaka Hannyaharamitu, E<lt>makamaka[at]cpan.orgE<gt>
 
 This library is free software; you can redistribute it and/or modify
 it under the same terms as Perl itself.